pluto/requirements.txt
Django==3.2.10
axblack==20210720
djangorestframework==3.12.4
pytz==2021.3